requirements.txt
# FILE: backend/requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis[asyncio]==5.0.3

# AI / Face Recognition
insightface==0.7.3
onnxruntime-gpu==1.17.3  # Use onnxruntime for CPU-only deployments
faiss-gpu==1.7.4         # Use faiss-cpu for CPU-only
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Settings
pydantic-settings==2.2.1
pydantic[email]==2.7.1

# HTTP Client
httpx==0.27.0

# Monitoring
prometheus-fastapi-instrumentator==7.0.0

# Utils
python-dotenv==1.0.1
aiofiles==23.2.1
